backed out r108916
svn:sync-xref-src-repo-rev: r108918

+++ src/VBox/Main/include/HGCMObjects.h
@@ -1,4 +1,4 @@
-/* $Id: HGCMObjects.h 62370 2016-07-20 17:12:05Z noreply@oracle.com $ */
+/* $Id: HGCMObjects.h 62372 2016-07-20 17:25:41Z noreply@oracle.com $ */
 /** @file
  *
  * HGCMObjects - Host-Guest Communication Manager objects header.
@@ -89,7 +89,7 @@ class HGCMObject
 
         uint32_t Handle()
         {
-            return (uint32_t)m_core.AvlCore.Key;
+            return m_core.AvlCore.Key;
         };
 
         HGCMOBJ_TYPE Type()
